feat: Add persist option
This prevents hyprland or another program from causing hyprdim to stop
working if the dim_inactive variable is changed.

Although it's possible to restore the previous state when all dim
threads are finished executing, there doesn't seem to be a compelling
reason to do this since hyprdim listens for window events anyway, which
would negate the point of having dim_inactive redefined as false.